fix migrate

+++ database/migrations/2023_12_28_022549_create_exchange_rates_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('exchange_rates', function (Blueprint $table) {
             $table->id();
-            $table->float('jpy');
-            $table->float('usd');
+            $table->integer('jpy');
+            $table->integer('usd');
             $table->string('exchange_rate_month', 100)->nullable();
             $table->timestamps();
             $table->softDeletes();
